Converted into a composer based project for writing unit tests.
This could've been done even with a single PHP file as index.php, but PHP unit testing can be done with PHPUnit package.
Composer is a package manager PHP uses, and it is far easier to set it up than try to install and load dependencies for PHPUnit on our own.

For this change, I had to remove the index.php and the readme file, so composer setup will run.