Enabled PSR-4 autoloading and created a validator class to test autoloading
When we want to write application for the long haul, it is always best to configure a standard autoloading scheme such as PSR-4.
This will let us keep/modify our directory structure for classes, without having to rewrite every file that uses a class that's changed.

This also makes easier when writing tests, we don't have to trace back and load every dependency, because autoload will take care of the deps.